+++ buffer/buffer_getline.c
@@ -1,5 +1,5 @@
 #include <buffer.h>
 
-int buffer_getline(buffer* b,char* x,unsigned long int len) {
+ssize_t buffer_getline(buffer* b,char* x,size_t len) {
   return buffer_get_token(b,x,len,"\n",1);
 }
