+++ buffer/buffer_init_free.c
@@ -1,7 +1,7 @@
 #include "buffer.h"
 
-void buffer_init_free(buffer* b,int (*op)(),int fd,
-		 char* y,unsigned long int ylen) {
+void buffer_init_free(buffer* b,ssize_t (*op)(),int fd,
+		 char* y,size_t ylen) {
   buffer_init(b,op,fd,y,ylen);
   b->todo=FREE;
 }
