+++ buffer/buffer_putalign.c
@@ -1,7 +1,7 @@
 #include "byte.h"
 #include "buffer.h"
 
-int buffer_putalign(buffer* b,const char* buf,unsigned long int len) {
+int buffer_putalign(buffer* b,const char* buf,size_t len) {
   unsigned int tmp;
   while (len>(tmp=b->a-b->p)) {
     byte_copy(b->x+b->p, tmp, buf);
