+++ byte/byte_diff.c
@@ -5,7 +5,7 @@
  * than, equal to, or greater than the string one[0], one[1], ...,
  * one[len-1]. When the strings are different, byte_diff does not read
  * bytes past the first difference. */
-int byte_diff(const void* a, unsigned long len, const void* b) {
+int byte_diff(const void* a, size_t len, const void* b) {
   register const unsigned char* s=a;
   register const unsigned char* t=b;
   register const unsigned char* u=t+len;
