+++ case/case_lowerb.c
@@ -1,6 +1,6 @@
 #include "case.h"
 
-void case_lowerb(void *S,unsigned long len) {
+void case_lowerb(void *S,size_t len) {
   char* s=(char*)S;
   unsigned char x;
   while (len > 0) {
