+++ fmt/fmt_fill.c
@@ -4,7 +4,7 @@
  * append padlen-srclen spaces after dest, if that is >= 0.  Truncate
  * only if total length is larger than maxlen.  Return number of
  * characters written. */
-unsigned long fmt_fill(char* dest,unsigned long srclen,unsigned long padlen,unsigned long maxlen) {
+size_t fmt_fill(char* dest,size_t srclen,size_t padlen,size_t maxlen) {
   long todo;
   char* olddest=dest;
   char* max=dest+maxlen;
