+++ fmt/fmt_longlong.c
@@ -1,6 +1,6 @@
 #include "fmt.h"
 
-unsigned int fmt_longlong(char *dest,signed long long int i) {
+size_t fmt_longlong(char *dest,signed long long int i) {
   if (i<0) {
     if (dest) *dest++='-';
     return fmt_ulonglong(dest,-i)+1;
