+++ fmt/fmt_strn.c
@@ -1,6 +1,6 @@
 #include "fmt.h"
 
-unsigned long fmt_strn(char *out,const char *in,unsigned long limit) {
+size_t fmt_strn(char *out,const char *in,size_t limit) {
   register char* s=out;
   register const char* t=in;
   register const char* u=in+limit;
