+++ fmt/fmt_ulong.c
@@ -1,6 +1,6 @@
 #include "fmt.h"
 
-unsigned int fmt_ulong(char *dest,unsigned long i) {
+size_t fmt_ulong(char *dest,unsigned long i) {
   register unsigned long len,tmp,len2;
   /* first count the number of bytes needed */
   for (len=1, tmp=i; tmp>9; ++len) tmp/=10;
