+++ fmt/fmt_xlong.c
@@ -5,7 +5,7 @@ static inline char tohex(char c) {
   return c>=10?c-10+'a':c+'0';
 }
 
-unsigned int fmt_xlong(char *dest,unsigned long i) {
+size_t fmt_xlong(char *dest,unsigned long i) {
   register unsigned long len,tmp;
   /* first count the number of bytes needed */
   for (len=1, tmp=i; tmp>15; ++len) tmp>>=4;
