+++ mmap.h
@@ -1,20 +1,22 @@
 #ifndef MMAP_H
 #define MMAP_H
 
+#include <stddef.h>
+
 /* open file for reading, mmap whole file, close file, write length of
  * map in filesize and return pointer to map. */
-char* mmap_read(const char *filename,unsigned long* filesize);
+char* mmap_read(const char *filename,size_t* filesize);
 
 /* open file for writing, mmap whole file privately (copy on write),
  * close file, write length of map in filesize and return pointer to
  * map. */
-char* mmap_private(const char *filename,unsigned long* filesize);
+char* mmap_private(const char *filename,size_t* filesize);
 
 /* open file for writing, mmap whole file shared, close file, write
  * length of map in filesize and return pointer to map. */
-char* mmap_shared(const char *filename,unsigned long* filesize);
+char* mmap_shared(const char *filename,size_t* filesize);
 
 /* unmap a mapped region */
-int mmap_unmap(char* mapped,unsigned long maplen);
+int mmap_unmap(char* mapped,size_t maplen);
 
 #endif
