+++ openreadclose.h
@@ -3,6 +3,6 @@
 
 #include "stralloc.h"
 
-int openreadclose(const char *filename,stralloc *buf,unsigned int initiallength);
+int openreadclose(const char *filename,stralloc *buf,size_t initiallength);
 
 #endif
