+++ readclose.h
@@ -3,7 +3,7 @@
 
 #include "stralloc.h"
 
-int readclose_append(int fd,stralloc *buf,unsigned int initlen);
-int readclose(int fd,stralloc *buf,unsigned int initlen);
+int readclose_append(int fd,stralloc *buf,size_t initlen);
+int readclose(int fd,stralloc *buf,size_t initlen);
 
 #endif
