+++ scan/scan_8long.c
@@ -1,6 +1,6 @@
 #include "scan.h"
 
-unsigned int scan_8long(const char *src,unsigned long *dest) {
+size_t scan_8long(const char *src,unsigned long *dest) {
   register const char *tmp=src;
   register unsigned long l=0;
   register unsigned char c;
