+++ scan/scan_xint.c
@@ -1,6 +1,6 @@
 #include "scan.h"
 
-unsigned int scan_xint(const char* src,unsigned int* dest) {
+size_t scan_xint(const char* src,unsigned int* dest) {
   unsigned long l;
   register int len=scan_xlong(src,&l);
   *dest=l;
